+++ packages/core/src/app/address/AddressForm.tsx
@@ -71,13 +71,9 @@ const AddressForm: React.FC<AddressFormProps> = ({
         getProviderWithCustomCheckout(config?.checkoutSettings.providerWithCustomCheckout),
     );
     // PayPal Fastlane stores keep the legacy phone input for now, due to incident
-    const isNewPhoneValidationExperimentEnabled =
+    const isPhoneNumberValidationEnabled =
         !isPayPalFastlaneEnabled &&
-        isExperimentEnabled(
-            config?.checkoutSettings,
-            'CHECKOUT-9019.use_new_phone_number_validation',
-            false,
-        );
+        (config?.checkoutSettings.isPhoneNumberValidationEnabled ?? false);
     const isNewGooglePlacesApiEnabled = isExperimentEnabled(
         config?.checkoutSettings,
         'CHECKOUT-10026.new_google_places_api',
@@ -236,9 +232,7 @@ const AddressForm: React.FC<AddressFormProps> = ({
                                 inputId={getAddressFormFieldInputId(addressFieldName)}
                                 // stateOrProvince can sometimes be a dropdown or input, so relying on id is not sufficient
                                 isFloatingLabelEnabled={isFloatingLabelEnabledValue}
-                                isNewPhoneValidationExperimentEnabled={
-                                    isNewPhoneValidationExperimentEnabled
-                                }
+                                isPhoneNumberValidationEnabled={isPhoneNumberValidationEnabled}
                                 key={`${field.id}-${field.name}`}
                                 label={
                                     field.custom || isExtraField(field) ? (