+++ packages/core/src/app/orderComments/OrderComments.tsx
@@ -1,17 +1,23 @@
 import { type FieldProps } from 'formik';
 import React, { type FunctionComponent, useCallback, useMemo } from 'react';
 
+import { useThemeContext } from '@bigcommerce/checkout/contexts';
 import { TranslatedString } from '@bigcommerce/checkout/locale';
 import { Fieldset, FormField, Label, Legend, TextInput } from '@bigcommerce/checkout/ui';
 
 const OrderComments: FunctionComponent = () => {
+    const { enhancedThemeV1 } = useThemeContext();
+    const orderCommentLabelId = enhancedThemeV1
+        ? 'shipping.order_comment_label_v2'
+        : 'shipping.order_comment_label';
+
     const renderLabel = useCallback(
         (name: string) => (
             <Label hidden htmlFor={name}>
-                <TranslatedString id="shipping.order_comment_label" />
+                <TranslatedString id={orderCommentLabelId} />
             </Label>
         ),
-        [],
+        [orderCommentLabelId],
     );
 
     const renderInput = useCallback(
@@ -24,10 +30,10 @@ const OrderComments: FunctionComponent = () => {
     const legend = useMemo(
         () => (
             <Legend>
-                <TranslatedString id="shipping.order_comment_label" />
+                <TranslatedString id={orderCommentLabelId} />
             </Legend>
         ),
-        [],
+        [orderCommentLabelId],
     );
 
     return (